(#864) Handle corrupt LiteDatabase files
If one of the LiteDatabase files is corrupted or empty or just not a
LiteDatabase file, the initialization of the app fails. This will
naively ignore the failure of the global database and allow it to act
as a read only of the default values. If the user database is corrupt,
we just throw a generic exception.

Obviously the generic exception is not ideal, but it allows to see
exactly what is happening.